requirements.txt
pywebview>=4.0
faster-whisper>=0.9.0
pydub>=0.25.1
